fix(PageHeader): restrict height of description block
